+++ dynamic-mapper-extension/src/main/java/dynamic/mapper/processor/extension/external/ProcessorExtensionCustomAlarm.java
@@ -40,9 +40,8 @@ import dynamic.mapper.processor.extension.ProcessorExtensionSource;
 import dynamic.mapper.processor.extension.ProcessorExtensionTarget;
 import dynamic.mapper.processor.model.C8YRequest;
 import dynamic.mapper.processor.model.ProcessingContext;
+import dynamic.mapper.processor.model.ProcessingException;
 import dynamic.mapper.processor.model.RepairStrategy;
-import dynamic.mapper.processor.ProcessingException;
-
 import dynamic.mapper.core.C8YAgent;
 import lombok.extern.slf4j.Slf4j;
 import org.joda.time.DateTime;
