+++ dynamic-mapper-service/src/main/java/dynamic/mapper/processor/extension/ProcessorExtensionSource.java
@@ -21,8 +21,9 @@
 
 package dynamic.mapper.processor.extension;
 
-import dynamic.mapper.processor.ProcessingException;
 import dynamic.mapper.processor.model.ProcessingContext;
+import dynamic.mapper.processor.model.ProcessingException;
+
 import org.springframework.stereotype.Component;
 
 @Component
