+++ dynamic-mapper-service/src/main/java/dynamic/mapper/processor/inbound/JSONProcessorInbound.java
@@ -43,8 +43,8 @@ import dynamic.mapper.model.Substitution;
 import dynamic.mapper.processor.model.SubstituteValue.TYPE;
 import dynamic.mapper.processor.model.SubstitutionEvaluation;
 import dynamic.mapper.processor.model.SubstituteValue;
-import dynamic.mapper.processor.ProcessingException;
 import dynamic.mapper.processor.model.ProcessingContext;
+import dynamic.mapper.processor.model.ProcessingException;
 import dynamic.mapper.processor.model.RepairStrategy;
 import lombok.extern.slf4j.Slf4j;
