+++ dynamic-mapper-service/src/main/java/dynamic/mapper/processor/outbound/JSONProcessorOutbound.java
@@ -33,8 +33,8 @@ import dynamic.mapper.connector.core.client.AConnectorClient;
 import dynamic.mapper.core.ConfigurationRegistry;
 import dynamic.mapper.model.Mapping;
 import dynamic.mapper.model.Substitution;
-import dynamic.mapper.processor.ProcessingException;
 import dynamic.mapper.processor.model.ProcessingContext;
+import dynamic.mapper.processor.model.ProcessingException;
 import dynamic.mapper.processor.model.SubstituteValue;
 import lombok.extern.slf4j.Slf4j;
