+++ dynamic-mapper-service/src/main/java/dynamic/mapper/service/NotificationSubscriptionService.java
@@ -32,7 +32,6 @@ import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
